pyproject.toml
[project]
name = "a2a-document-creator"
version = "0.1.0"
description = "A2A Multi-Format Document Creation Agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "python-dotenv>=1.1.0",
    "langchain-google-genai",
    "uvicorn>=0.34.2",
    "python-pptx>=1.0.0",
    "python-docx>=1.1.0",
    "matplotlib>=3.8.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"